Replace flaky minifb crate with winit/pixels crates.
Background: minifb was causing issues when resizing the window and the cocoa
event pump would frequently get clogged, causing code in the minifb library to
block and slowing down the emulator. Moved to winit/pixels to manage the window
manually and gain more control over the low-level window events. The winit crate
seems to handle window events in cocoa much better. It does not block the main
loop on window resizing and purges the event queue much more effectively.